fix: Presentation Layer 날씨 Model 최저기온, 최고기온 정수형으로 변환

+++ app/src/main/java/com/sangmeebee/weatherlist/model/WeatherModel.kt
@@ -11,8 +11,8 @@ sealed class WeatherModel(open val type: WeatherViewType) {
         override val type: WeatherViewType,
         val city: String,
         val date: String,
-        val tempMin: Double,
-        val tempMax: Double,
+        val tempMin: Int,
+        val tempMax: Int,
         val iconName: String,
         val iconType: String,
     ) : WeatherModel(type)
